refactor: fix clang-tidy warnings

+++ kernel/include/memlayout.h
@@ -5,7 +5,7 @@
 #include "types.h"
 
 /*
- * 0x0F0XXXXX   free bottom top
+ * 0x0F0XXXXX   free mem top
  * ....
  * ....
  * ....       + MEMORY SIZE - kernel size
@@ -31,7 +31,7 @@ extern symbol __kernel_text_end;
 #define __KERNEL_END	  ((usize)__kernel_end)
 #define __KERNEL_TEXT_END ((usize)__kernel_text_end)
 
-#define MB_TO_BYTES(MB) ((MB) * 1000 * 1000)
+#define MB_TO_BYTES(MB) ((MB) * (usize)1000 * (usize)1000)
 #define KERNEL_BOTTOM	(__KERNEL_START)
 #define KERNEL_TOP	(__KERNEL_END)
 #define FREE_MEM_BOTTOM (KERNEL_TOP)
